+++ src/includes/restApi/SettingsController.php
@@ -2,7 +2,7 @@
 /**
  * Plugin settings endpoint for WordPress REST API.
  *
- * @link       https://t.me/manzoorwanijk
+ * @link       https://manzoorwani.dev
  * @since      1.7.0
  *
  * @package    WPTelegram_Widget
@@ -68,20 +68,20 @@ class SettingsController extends RESTController {
 		register_rest_route(
 			self::NAMESPACE,
 			self::REST_BASE,
-			array(
-				array(
+			[
+				[
 					'methods'             => WP_REST_Server::READABLE,
-					'callback'            => array( $this, 'get_settings' ),
-					'permission_callback' => array( $this, 'settings_permissions' ),
+					'callback'            => [ $this, 'get_settings' ],
+					'permission_callback' => [ $this, 'settings_permissions' ],
 					'args'                => self::get_settings_params( 'view' ),
-				),
-				array(
+				],
+				[
 					'methods'             => WP_REST_Server::CREATABLE,
-					'callback'            => array( $this, 'update_settings' ),
-					'permission_callback' => array( $this, 'settings_permissions' ),
+					'callback'            => [ $this, 'update_settings' ],
+					'permission_callback' => [ $this, 'settings_permissions' ],
 					'args'                => self::get_settings_params( 'edit' ),
-				),
-			)
+				],
+			]
 		);
 	}
 
@@ -105,21 +105,51 @@ class SettingsController extends RESTController {
 
 		$settings = WPTG_Widget()->options()->get_data();
 
-		// If we have somethings saved.
-		if ( ! empty( $settings ) ) {
-			unset( $settings['messages'] );
-			unset( $settings['last_update_id'] );
-			return $settings;
+		// If we have something saved.
+		if ( ! isset( $settings['ajax_widget'] ) ) {
+			return self::get_default_values();
 		}
 
-		// Get the default values.
-		$settings = self::get_settings_params();
-
-		foreach ( $settings as $key => $args ) {
-			$settings[ $key ] = isset( $args['default'] ) ? $args['default'] : '';
-		}
+		unset( $settings['messages'] );
+		unset( $settings['last_update_id'] );
 
 		return $settings;
+	}
+
+	/**
+	 * Get the default values for settings.
+	 *
+	 * @return array
+	 */
+	public static function get_default_values() {
+		return [
+			'ajax_widget'   => [
+				'username' => '',
+				'width'    => '100%',
+				'height'   => '600px',
+			],
+			'legacy_widget' => [
+				'username'     => '',
+				'width'        => '100%',
+				'height'       => '600px',
+				'bot_token'    => '',
+				'author_photo' => 'auto',
+				'num_messages' => '5',
+			],
+			'join_link'     => [
+				'url'        => '',
+				'text'       => '',
+				'bgcolor'    => '#389ce9',
+				'text_color' => '#fff',
+				'post_types' => [ 'post' ],
+				'position'   => 'after_content',
+				'priority'   => '10',
+			],
+			'advanced'      => [
+				'telegram_blocked'  => false,
+				'google_script_url' => '',
+			],
+		];
 	}
 
 	/**
@@ -168,99 +198,108 @@ class SettingsController extends RESTController {
 	 * @return array Query parameters for the settings.
 	 */
 	public static function get_settings_params( $context = 'edit' ) {
-		$ajax_widget_props = array(
-			'username' => array(
-				'type'     => 'string',
-				'required' => ( 'edit' === $context ),
-				'pattern'  => Utils::enhance_regex(
-					self::TG_USERNAME_PATTERN,
-					true
-				),
-			),
-			'width'    => array(
+		$ajax_widget_props = [
+			'username' => [
 				'type'    => 'string',
-				'default' => '100%',
-			),
-			'height'   => array(
-				'type'    => 'string',
-				'default' => '600px',
-			),
-		);
+				'pattern' => Utils::enhance_regex( self::TG_USERNAME_PATTERN, true ),
+			],
+			'width'    => [
+				'type' => 'string',
+			],
+			'height'   => [
+				'type' => 'string',
+			],
+		];
 
-		return array(
-			'ajax_widget'   => array(
-				'type'       => 'object',
-				'properties' => $ajax_widget_props,
-			),
-			'legacy_widget' => array(
-				'type'       => 'object',
-				'properties' => array_merge(
+		return [
+			'ajax_widget'   => [
+				'type'              => 'object',
+				'properties'        => $ajax_widget_props,
+				'sanitize_callback' => [ __CLASS__, 'sanitize_param' ],
+				'validate_callback' => 'rest_validate_request_arg',
+			],
+			'legacy_widget' => [
+				'type'              => 'object',
+				'sanitize_callback' => [ __CLASS__, 'sanitize_param' ],
+				'validate_callback' => 'rest_validate_request_arg',
+				'properties'        => array_merge(
 					$ajax_widget_props,
-					array(
-						'username'     => array_merge(
-							$ajax_widget_props['username'],
-							array( 'required' => false )
-						),
-						'bot_token'    => array(
+					[
+						'bot_token'    => [
 							'type'    => 'string',
-							'pattern' => Utils::enhance_regex(
-								API::BOT_TOKEN_PATTERN,
-								true
-							),
-						),
-						'author_photo' => array(
-							'type'    => 'string',
-							'default' => 'auto',
-							'enum'    => array( 'auto', 'always_show', 'always_hide' ),
-						),
-						'num_messages' => array(
-							'type'    => 'string',
-							'default' => '5',
-						),
-					)
-				),
-			),
-			'join_link'     => array(
-				'type'       => 'object',
-				'properties' => array(
-					'url'        => array(
-						'type'   => 'string',
-						'format' => 'uri',
-					),
-					'text'       => array(
-						'type' => 'string',
-					),
-					'post_types' => array(
-						'type'    => 'array',
-						'default' => array( 'post' ),
-						'items'   => array(
+							'pattern' => Utils::enhance_regex( API::BOT_TOKEN_PATTERN, true ),
+						],
+						'author_photo' => [
 							'type' => 'string',
-						),
-					),
-					'position'   => array(
-						'type'    => 'string',
-						'default' => 'after_content',
-						'enum'    => array( 'before_content', 'after_content' ),
-					),
-					'priority'   => array(
-						'type'    => 'string',
-						'default' => '10',
-					),
+							'enum' => [ 'auto', 'always_show', 'always_hide' ],
+						],
+						'num_messages' => [
+							'type' => 'string',
+						],
+					]
 				),
-			),
-			'advanced'      => array(
-				'type'       => 'object',
-				'properties' => array(
-					'telegram_blocked'  => array(
-						'type'    => 'boolean',
-						'default' => false,
-					),
-					'google_script_url' => array(
+			],
+			'join_link'     => [
+				'type'              => 'object',
+				'sanitize_callback' => [ __CLASS__, 'sanitize_param' ],
+				'validate_callback' => 'rest_validate_request_arg',
+				'properties'        => [
+					'url'        => [
 						'type'   => 'string',
 						'format' => 'uri',
-					),
-				),
-			),
-		);
+					],
+					'text'       => [
+						'type' => 'string',
+					],
+					'bgcolor'    => [
+						'type' => 'string',
+					],
+					'text_color' => [
+						'type' => 'string',
+					],
+					'post_types' => [
+						'type'  => 'array',
+						'items' => [
+							'type' => 'string',
+						],
+					],
+					'position'   => [
+						'type' => 'string',
+						'enum' => [ 'before_content', 'after_content' ],
+					],
+					'priority'   => [
+						'type' => 'string',
+					],
+				],
+			],
+			'advanced'      => [
+				'type'              => 'object',
+				'sanitize_callback' => [ __CLASS__, 'sanitize_param' ],
+				'validate_callback' => 'rest_validate_request_arg',
+				'properties'        => [
+					'telegram_blocked'  => [
+						'type' => 'boolean',
+					],
+					'google_script_url' => [
+						'type'   => 'string',
+						'format' => 'uri',
+					],
+				],
+			],
+		];
+	}
+
+	/**
+	 * Sanitize the request param.
+	 *
+	 * @since x.y.z
+	 *
+	 * @param mixed           $value   Value of the param.
+	 * @param WP_REST_Request $request WP REST API request.
+	 * @param string          $param     The param key.
+	 */
+	public static function sanitize_param( $value, WP_REST_Request $request, $param ) {
+		// Lets make the value safer.
+		return Utils::sanitize( $value );
 	}
 }